Restructured player for more DRY code and changed checkpoint collider
