+++ Assets/Prefabs/Player/Player.prefab
@@ -857,6 +857,7 @@ GameObject:
   m_Component:
   - component: {fileID: 6554401562002261475}
   - component: {fileID: 1152011476486908822}
+  - component: {fileID: 2213690685019397990}
   - component: {fileID: 3837601306821772295}
   - component: {fileID: 5284907276310118903}
   - component: {fileID: 2281859966435908982}
@@ -902,6 +903,18 @@ Rigidbody:
   m_Interpolate: 0
   m_Constraints: 112
   m_CollisionDetection: 2
+--- !u!114 &2213690685019397990
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6554401562002261474}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: e552000049255dd43bb59d987d955d76, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!114 &3837601306821772295
 MonoBehaviour:
   m_ObjectHideFlags: 0
