+++ Assets/Prefabs/RaceComponents/Checkpoint.prefab
@@ -11,8 +11,8 @@ GameObject:
   - component: {fileID: 6929813948986943113}
   - component: {fileID: 6604599465974141602}
   - component: {fileID: 45514448837263381}
-  - component: {fileID: 8225794825143513771}
   - component: {fileID: 4528755849620568021}
+  - component: {fileID: 4315410451385234515}
   m_Layer: 2
   m_Name: Checkpoint
   m_TagString: Checkpoint
@@ -85,20 +85,6 @@ MeshRenderer:
   m_SortingLayer: 0
   m_SortingOrder: 0
   m_AdditionalVertexStreams: {fileID: 0}
---- !u!136 &8225794825143513771
-CapsuleCollider:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 243084638427717879}
-  m_Material: {fileID: 0}
-  m_IsTrigger: 1
-  m_Enabled: 1
-  m_Radius: 0.5000001
-  m_Height: 2
-  m_Direction: 1
-  m_Center: {x: 0.000000059604645, y: 0, z: -0.00000008940697}
 --- !u!114 &4528755849620568021
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -111,3 +97,21 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 464d2a81658732443b88cac478983f59, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  checkpoint: 0
+  finish: 0
+  raceController: {fileID: 0}
+  passed: 0
+--- !u!64 &4315410451385234515
+MeshCollider:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 243084638427717879}
+  m_Material: {fileID: 0}
+  m_IsTrigger: 1
+  m_Enabled: 1
+  serializedVersion: 4
+  m_Convex: 1
+  m_CookingOptions: 30
+  m_Mesh: {fileID: 10206, guid: 0000000000000000e000000000000000, type: 0}
