+++ Assets/Scenes/SampleScene.unity
@@ -544,7 +544,7 @@ Transform:
   - {fileID: 1451997196}
   - {fileID: 1487787894}
   m_Father: {fileID: 0}
-  m_RootOrder: 2
+  m_RootOrder: 3
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1001 &689847433
 PrefabInstance:
@@ -1510,7 +1510,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 6929813948986943113, guid: 5d5bfd8288ae1fb4b9ae925ac5342150, type: 3}
       propertyPath: m_RootOrder
-      value: 3
+      value: 2
       objectReference: {fileID: 0}
     - target: {fileID: 6929813948986943113, guid: 5d5bfd8288ae1fb4b9ae925ac5342150, type: 3}
       propertyPath: m_LocalPosition.x
