+++ Assets/Scripts/Player/InitPlayer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e552000049255dd43bb59d987d955d76
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
